requirements.txt
slackclient==2.9.4
python-dotenv==0.20.0